Update AmazonS3FileStorage.java
Is there some reason the FileDescriptor.id field (a UUID) is used to store the file on amazon? I understand that somehow the file name may be blank, but it is a required field for FileDescriptor. I modified the getFileName() to use FileDescriptor.name as long as it is not blank. If it is, resort to  the UUID.